+++ CHANGELOG.md
@@ -1,1 +1,5 @@
 ## Changelog
+- Add Circle Alert for Orbs&Bombs on Coiled Altar
+- Fix Shield Alert for P3
+- use castBarID for interrupt alert
+- update pace data for twin fangs and coiled altar